Move datastore classes to separate sub library.
