+++ searchlib/src/vespa/searchlib/attribute/enumstore.h
@@ -9,7 +9,7 @@
 #include <vespa/vespalib/util/array.h>
 #include <vespa/vespalib/util/stringfmt.h>
 #include <cmath>
-#include <vespa/searchlib/btree/entryref.h>
+#include <vespa/searchlib/datastore/entryref.h>
 #include <vespa/searchlib/btree/btreenode.h>
 #include <vespa/searchlib/btree/btreenodeallocator.h>
 #include <vespa/searchlib/btree/btree.h>
