+++ searchlib/src/vespa/searchlib/attribute/enumstorebase.h
@@ -4,7 +4,7 @@
 
 #include "address_space.h"
 #include <vespa/searchcommon/attribute/iattributevector.h>
-#include <vespa/searchlib/btree/datastore.h>
+#include <vespa/searchlib/datastore/datastore.h>
 #include <vespa/searchlib/util/memoryusage.h>
 #include <vespa/vespalib/stllike/asciistream.h>
 #include <vespa/vespalib/util/array.h>
