+++ searchlib/src/vespa/searchlib/attribute/loadedvalue.h
@@ -4,7 +4,7 @@
 
 #include <vespa/searchcommon/common/undefinedvalues.h>
 #include <vespa/searchlib/attribute/enumstorebase.h>
-#include <vespa/searchlib/btree/entryref.h>
+#include <vespa/searchlib/datastore/entryref.h>
 
 namespace search
 {
