+++ searchlib/src/vespa/searchlib/btree/btreenode.cpp
@@ -1,5 +1,6 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
+#include <vespa/fastos/fastos.h>
 #include "btreenode.h"
 #include "btreenode.hpp"
