+++ searchlib/src/vespa/searchlib/btree/btreenodestore.cpp
@@ -1,11 +1,12 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
+#include <vespa/fastos/fastos.h>
 #include "btreenodestore.hpp"
-#include "datastore.h"
 #include "btreenode.h"
 #include "btreerootbase.h"
 #include "btreeroot.h"
 #include "btreenodeallocator.h"
+#include <vespa/searchlib/datastore/datastore.h>
 
 namespace search
 {
