+++ searchlib/src/vespa/searchlib/btree/btreestore.cpp
@@ -1,13 +1,9 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
+#include <vespa/fastos/fastos.h>
 #include "btreestore.h"
-#include "datastore.h"
-#include "btreenode.h"
-#include "btreerootbase.h"
-#include "btreeroot.h"
-#include "btreenodeallocator.h"
-#include "btreeiterator.hpp"
 #include "btreestore.hpp"
+#include "btreeiterator.hpp"
 
 namespace search
 {
