+++ searchlib/src/vespa/searchlib/datastore/CMakeLists.txt
@@ -1,0 +1,8 @@
+# Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+vespa_add_library(searchlib_datastore OBJECT
+    SOURCES
+    bufferstate.cpp
+    datastore.cpp
+    datastorebase.cpp
+    DEPENDS
+)
