+++ searchlib/src/vespa/searchlib/datastore/bufferstate.cpp
