+++ searchlib/src/vespa/searchlib/datastore/bufferstate.h
