+++ searchlib/src/vespa/searchlib/datastore/datastore.h
@@ -3,7 +3,6 @@
 #pragma once
 
 #include "datastorebase.h"
-#include "btreenode.h"
 
 namespace search
 {
