+++ searchlib/src/vespa/searchlib/datastore/datastorebase.cpp
