+++ searchlib/src/vespa/searchlib/datastore/entryref.h
