+++ searchlib/src/vespa/searchlib/memoryindex/compact_document_words_store.h
@@ -1,8 +1,8 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 #pragma once
 
-#include <vespa/searchlib/btree/datastore.h>
-#include <vespa/searchlib/btree/entryref.h>
+#include <vespa/searchlib/datastore/datastore.h>
+#include <vespa/searchlib/datastore/entryref.h>
 #include <vespa/searchlib/util/memoryusage.h>
 #include <vespa/vespalib/util/array.h>
 #include <vespa/vespalib/stllike/hash_map.h>
