+++ searchlib/src/vespa/searchlib/memoryindex/featurestore.h
@@ -2,7 +2,7 @@
 
 #pragma once
 
-#include <vespa/searchlib/btree/datastore.h>
+#include <vespa/searchlib/datastore/datastore.h>
 #include <vespa/searchlib/index/docidandfeatures.h>
 #include <vespa/searchlib/bitcompression/compression.h>
 #include <vespa/searchlib/bitcompression/posocccompression.h>
