+++ searchlib/src/vespa/searchlib/memoryindex/i_document_insert_listener.h
@@ -1,6 +1,6 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 #pragma once
-#include <vespa/searchlib/btree/entryref.h>
+#include <vespa/searchlib/datastore/entryref.h>
 
 namespace search {
 namespace memoryindex {
