+++ searchlib/src/vespa/searchlib/tensor/dense_tensor_store.cpp
@@ -11,7 +11,7 @@
 #include <vespa/vespalib/util/stringfmt.h>
 #include <vespa/vespalib/util/macro.h>
 #include <vespa/document/util/serializable.h>
-#include <vespa/searchlib/btree/datastore.hpp>
+#include <vespa/searchlib/datastore/datastore.hpp>
 
 using vespalib::tensor::Tensor;
 using vespalib::tensor::DenseTensor;
